+++ changelog.md
@@ -18,7 +18,8 @@
 
 * All Views
   * The `modelEvents` and `collectionEvents` can now have a function configured as the value in the `{ "event:name": "value" }` configuration hash
-
+  * A view that uses `bindTo` for its own "close" event will have it's close handler called correctly
+  * Returning `false` from the `onBeforeClose` method will prevent the view from being closed
 
 ### v1.0.0-beta6 [view commit logs](https://github.com/marionettejs/backbone.marionette/compare/v1.0.0-beta5...v1.0.0-beta6)